pyproject.toml
[tool.poetry]
name = "forex"
version = "0.1.0"
description = "Forex rates application"
authors = ["sdediego <sergiodediego@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
psycopg2 = "^2.9.1"
djangorestframework = "^3.12.4"
marshmallow = "^3.13.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
